# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GradoEats Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: orders
  - name: drivers
  - name: dashboard
  - name: activities
  - name: alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_delivery, delivered, cancelled]
        - name: driverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Created driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a driver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /drivers/{id}/location:
    patch:
      operationId: updateDriverLocation
      tags: [drivers]
      summary: Update driver GPS location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverLocationBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /drivers/{id}/activities:
    get:
      operationId: getDriverActivities
      tags: [drivers]
      summary: Get activity timeline for a specific driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Driver activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  /drivers/{id}/today:
    get:
      operationId: getDriverTodayStats
      tags: [drivers]
      summary: Get today's stats for a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Today's stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverTodayStats"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List recent activities (all drivers)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: driverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List active alerts (waiting orders, inactive drivers)
      responses:
        "200":
          description: Active alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard summary stats
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/revenue:
    get:
      operationId: getRevenueStats
      tags: [dashboard]
      summary: Revenue stats grouped by day
      responses:
        "200":
          description: Revenue per day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDay"

  /dashboard/driver-stats:
    get:
      operationId: getDriverStats
      tags: [dashboard]
      summary: Stats per driver
      responses:
        "200":
          description: Per-driver stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, assigned, in_delivery, delivered, cancelled]
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerPhone
        - deliveryAddress
        - items
        - totalAmount
        - status
        - createdAt
        - updatedAt

    CreateOrderBody:
      type: object
      properties:
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: string
        totalAmount:
          type: number
        notes:
          type: string
        sourceUrl:
          type: string
      required:
        - orderNumber
        - customerName
        - customerPhone
        - deliveryAddress
        - items
        - totalAmount

    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, assigned, in_delivery, delivered, cancelled]
        driverId:
          type: integer
          nullable: true
        notes:
          type: string

    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        vehicleType:
          type: string
        status:
          type: string
          enum: [available, busy, offline]
        rating:
          type: number
        totalDeliveries:
          type: integer
        totalRevenue:
          type: number
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        lastActiveAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - vehicleType
        - status
        - rating
        - totalDeliveries
        - totalRevenue
        - createdAt

    CreateDriverBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        vehicleType:
          type: string
        avatarUrl:
          type: string
      required:
        - name
        - phone
        - vehicleType

    UpdateDriverBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        vehicleType:
          type: string
        status:
          type: string
          enum: [available, busy, offline]
        rating:
          type: number
        avatarUrl:
          type: string

    UpdateDriverLocationBody:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required:
        - lat
        - lng

    Activity:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        orderId:
          type: integer
          nullable: true
        orderNumber:
          type: string
          nullable: true
        action:
          type: string
          enum: [order_assigned, order_picked_up, order_delivered, order_cancelled, status_online, status_offline, status_available, location_updated]
        details:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - action
        - createdAt

    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [order_waiting_too_long, driver_inactive, driver_offline_with_order]
        severity:
          type: string
          enum: [warning, critical]
        message:
          type: string
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        orderId:
          type: integer
          nullable: true
        orderNumber:
          type: string
          nullable: true
        minutesElapsed:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - type
        - severity
        - message
        - createdAt

    DriverTodayStats:
      type: object
      properties:
        driverId:
          type: integer
        todayDeliveries:
          type: integer
        todayRevenue:
          type: number
        todayOnlineMinutes:
          type: integer
        activeOrderId:
          type: integer
          nullable: true
        activeOrderNumber:
          type: string
          nullable: true
        lastActivityAt:
          type: string
          nullable: true
      required:
        - driverId
        - todayDeliveries
        - todayRevenue
        - todayOnlineMinutes

    DashboardSummary:
      type: object
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        inDeliveryOrders:
          type: integer
        deliveredOrders:
          type: integer
        totalRevenue:
          type: number
        todayRevenue:
          type: number
        todayOrders:
          type: integer
        activeDrivers:
          type: integer
        averageRating:
          type: number
        alertCount:
          type: integer
      required:
        - totalOrders
        - pendingOrders
        - inDeliveryOrders
        - deliveredOrders
        - totalRevenue
        - todayRevenue
        - todayOrders
        - activeDrivers
        - averageRating
        - alertCount

    RevenueDay:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders

    DriverStats:
      type: object
      properties:
        driverId:
          type: integer
        driverName:
          type: string
        deliveries:
          type: integer
        revenue:
          type: number
        rating:
          type: number
        status:
          type: string
      required:
        - driverId
        - driverName
        - deliveries
        - revenue
        - rating
        - status