# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityAction } from "./activityAction";

export interface Activity {
  id: number;
  driverId?: number | null;
  driverName?: string | null;
  orderId?: number | null;
  orderNumber?: string | null;
  action: ActivityAction;
  details?: string | null;
  createdAt: string;
}