# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";
import type { AlertType } from "./alertType";

export interface Alert {
  id: string;
  type: AlertType;
  severity: AlertSeverity;
  message: string;
  driverId?: number | null;
  driverName?: string | null;
  orderId?: number | null;
  orderNumber?: string | null;
  minutesElapsed?: number | null;
  createdAt: string;
}