# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalOrders: number;
  pendingOrders: number;
  inDeliveryOrders: number;
  deliveredOrders: number;
  totalRevenue: number;
  todayRevenue: number;
  todayOrders: number;
  activeDrivers: number;
  averageRating: number;
  alertCount: number;
}