# lib/api-zod/src/generated/types/driver.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverStatus } from "./driverStatus";

export interface Driver {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  vehicleType: string;
  status: DriverStatus;
  rating: number;
  totalDeliveries: number;
  totalRevenue: number;
  lat?: number | null;
  lng?: number | null;
  avatarUrl?: string | null;
  lastActiveAt?: string | null;
  createdAt: string;
}