# lib/api-zod/src/generated/types/driverTodayStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverTodayStats {
  driverId: number;
  todayDeliveries: number;
  todayRevenue: number;
  todayOnlineMinutes: number;
  activeOrderId?: number | null;
  activeOrderNumber?: string | null;
  lastActivityAt?: string | null;
}